@@ -62,7 +62,7 @@ namespace Stride.Assets
             {
                 Name = PlatformType.Windows.ToString(),
                 IsAvailable = true,
-                TargetFramework = "net5.0-windows",
+                TargetFramework = "net6.0-windows",
                 RuntimeIdentifier = "win-x64",
                 Type = PlatformType.Windows
             };
@@ -121,7 +121,7 @@ namespace Stride.Assets
             {
                 Name = PlatformType.Linux.ToString(),
                 IsAvailable = true,
-                TargetFramework = "net5.0",
+                TargetFramework = "net6.0",
                 RuntimeIdentifier = "linux-x64",
                 Type = PlatformType.Linux,
             };
@@ -132,7 +132,7 @@ namespace Stride.Assets
             {
                 Name = PlatformType.macOS.ToString(),
                 IsAvailable = true,
-                TargetFramework = "net5.0",
+                TargetFramework = "net6.0",
                 RuntimeIdentifier = "osx-x64",
                 Type = PlatformType.macOS,
             };
